HH84: update student types
