+++ src/types/student-grades/student-grades.response.ts
@@ -1,6 +1,15 @@
-import { ImportDetails, ImportErrors } from './student-grades';
-
 export interface ImportResultResponse {
   added: ImportDetails;
   errors: ImportErrors;
+}
+
+export interface ImportDetails {
+  count: number;
+  details: string[];
+}
+
+export interface ImportErrors {
+  csvDuplicates: ImportDetails;
+  databaseDuplicates: ImportDetails;
+  invalidEmails: ImportDetails;
 }
