+++ src/types/student-profile/student-profile.ts
@@ -1,3 +1,5 @@
+import { ExpectedContractType, ExpectedTypeWork } from './student-profile.request';
+
 import { StudentEntity } from '../student';
 
 export interface StudentProfileEntity {
@@ -21,19 +23,4 @@ export interface StudentProfileEntity {
   student: StudentEntity;
   createdAt: Date;
   updatedAt: Date;
-}
-
-export enum ExpectedTypeWork {
-  ONSITE = 'onsite',
-  RELOCATION_READY = 'relocation_ready',
-  REMOTE_ONLY = 'remote_only',
-  HYBRID = 'hybrid',
-  NO_PREFERENCE = 'no_preference',
-}
-
-export enum ExpectedContractType {
-  UOP_ONLY = 'uop_only',
-  B2B_POSSIBLE = 'b2b_possible',
-  UZ_UOD_POSSIBLE = 'uz_uod_possible',
-  NO_PREFERENCE = 'no_preference',
 }
